requirements.txt
genlayer-testing>=0.1.0
pytest>=7.0.0
